[INJIMOB-1204]: Renaming sharing utils file and removing state variable in QRCodeOverlay
