+++ shared/sharing/sharingImageUtils.ts
